+++ tests/Unit/Methods/GlobalMethodsTest.php
@@ -3,24 +3,32 @@
 namespace SebastiaanLuca\Helpers\Tests\Unit\Methods;
 
 use Carbon\Carbon;
-use Laravelista\Ekko\Ekko;
 use SebastiaanLuca\Helpers\Methods\GlobalHelpersServiceProvider;
 use SebastiaanLuca\Helpers\Pipe\Item;
 use SebastiaanLuca\Helpers\Tests\TestCase;
 
 class GlobalMethodsTest extends TestCase
 {
-    public function test it returns a random bool()
+    /**
+     * @test
+     */
+    function it returns a random bool()
     {
         $this->assertInternalType('bool', rand_bool());
     }
 
-    public function test it wraps a string()
+    /**
+     * @test
+     */
+    function it wraps a string()
     {
         $this->assertSame('wrappermiddlewrapper', str_wrap('middle', 'wrapper'));
     }
 
-    public function test it detects an associative array()
+    /**
+     * @test
+     */
+    function it detects an associative array()
     {
         $this->assertTrue(is_assoc_array([
             'key1' => 'value1',
@@ -30,7 +38,10 @@ class GlobalMethodsTest extends TestCase
         $this->assertFalse(is_assoc_array(['value1', 'value2']));
     }
 
-    public function test it expands a flat array()
+    /**
+     * @test
+     */
+    function it expands a flat array()
     {
         $this->assertSame([
             'a' => [
@@ -47,13 +58,19 @@ class GlobalMethodsTest extends TestCase
         ]));
     }
 
-    public function test it removes values from an array()
+    /**
+     * @test
+     */
+    function it removes values from an array()
     {
         $this->assertSame(['a', 1], array_without(['a', 'b', 1], 'b'));
         $this->assertSame(['a'], array_without(['a', 'b', 1], ['b', 1]));
     }
 
-    public function test it pulls values from an array()
+    /**
+     * @test
+     */
+    function it pulls values from an array()
     {
         $array = ['a', 'b', 'c'];
 
@@ -61,7 +78,10 @@ class GlobalMethodsTest extends TestCase
         $this->assertSame(['a', 'c'], $array);
     }
 
-    public function test it pulls a value from an array()
+    /**
+     * @test
+     */
+    function it pulls a value from an array()
     {
         $array = ['a', 'b', 'c'];
 
@@ -69,7 +89,10 @@ class GlobalMethodsTest extends TestCase
         $this->assertSame(['a', 'c'], $array);
     }
 
-    public function test it generates an array hash()
+    /**
+     * @test
+     */
+    function it generates an array hash()
     {
         $hash = '9ae1f8db3c2cc8381e0811dda3316176';
 
@@ -77,7 +100,10 @@ class GlobalMethodsTest extends TestCase
         $this->assertNotSame($hash, array_hash(['value1', 'value2']));
     }
 
-    public function test it generate an object hash()
+    /**
+     * @test
+     */
+    function it generate an object hash()
     {
         $object = new \stdClass;
         $object->property = 'value';
@@ -91,21 +117,33 @@ class GlobalMethodsTest extends TestCase
         $this->assertNotSame($hash, object_hash($object));
     }
 
-    public function test it creates a carbon instance from a string()
+    /**
+     * @test
+     */
+    function it creates a carbon instance from a string()
     {
         $this->assertEquals(new Carbon('tomorrow'), carbonize('tomorrow'));
     }
 
-    public function test it creates a pipe item()
+    /**
+     * @test
+     */
+    function it creates a pipe item()
     {
         $this->assertInstanceOf(Item::class, take('value'));
     }
 
-    public function test it checks if a public method exists()
+    /**
+     * @test
+     */
+    function it checks if a public method exists()
     {
         $class = new class
         {
-            public function myMethod()
+            /**
+             * @test
+             */
+            function myMethod()
             {
                 return true;
             }
@@ -115,14 +153,20 @@ class GlobalMethodsTest extends TestCase
         $this->assertFalse(has_public_method($class, 'myInvalidMethod'));
     }
 
-    public function test it returns the current locale()
+    /**
+     * @test
+     */
+    function it returns the current locale()
     {
         config(['app.locale' => $locale = 'randomlocale']);
 
         $this->assertSame($locale, locale());
     }
 
-    public function test it returns the fallback locale if locale not set()
+    /**
+     * @test
+     */
+    function it returns the fallback locale if locale not set()
     {
         config(['app.locale' => null]);
         config(['app.fallback_locale' => $locale = 'fallbacklocale']);